Make "add from a web address" a shared part, for the tools that qualify
The picker can now bring a URL importer with it. One flag in a tool.toml turns
it on:

    [picker.urls]
    summary = "Add from a web address"
    button  = "Download images"
    noun    = "image"

That flag ships the module, its stylesheet and the widened img-src the page
needs, together, so the three cannot drift apart - and the build refuses to
finish if the panel is then left off body.html, because a page would otherwise
carry a network permission it never uses.

Nothing changes for a visitor today. Images to Video already had this, written
into its own src/remote.js and its own styles.css and its own [csp] table; it is
the same feature, in one place, ready for the next tool that can honestly take
it.

WHY ONLY ONE TOOL HAS IT

It was asked for on every tool by default, and it cannot be. The importer copies
each image through a <canvas>, so it arrives as a re-encoded JPEG rather than
the bytes the server sent. That is invisible in a tool whose pictures are headed
into a lossy video codec, and a lie in one that is going to tell you about the
bytes:

  * the EXIF viewer would show tags a canvas had already destroyed - the exact
    thing the tool exists to show;
  * the compressor would report its saving against a re-encode rather than
    against the file somebody actually has;
  * Images to PDF would decode the JPEGs it was just taught to copy through
    without decoding;
  * the video cropper needs the file's bytes for its demuxer, and a <video>
    element yields a playback stream, so the exact crop would silently drop to
    re-recording in real time.

So the test is not "would it work" but "would the answer still be true", and a
tool that would have to misdescribe what it was given goes without. Only Images
to Video passes today. The four claims of "no network feature of any kind" on
the other pages stay true, because they stay accurate.

WHY <img> AND NOT fetch()

Keeping the original bytes would need fetch(), which would need connect-src
opened to arbitrary origins - and the promise that there is nowhere for your
files to go would stop being true. Only img-src opens, and that is a one-way
door: pictures come in, data cannot go out. The re-encode is the price of
keeping the promise, and it is why the tools above have to decline rather than
accept a worse version.

Verified in a browser. On Images to Video: a malformed address is refused before
the network is touched, a real one downloads and lands in the list, the row
still says which host it came from, the textarea clears and the button comes
back. On the EXIF viewer: no panel, no module, the picker still works, and an
http image is still refused by the Content-Security-Policy. Every page still has
connect-src pointed at Google and nothing else, two builds are byte-identical,
and all 61 emitted scripts re-tokenise.
