+++ templates/partials/url-import.html
@@ -1,0 +1,33 @@
+<!--
+  "Add from a web address". Rendered only for a tool whose tool.toml sets
+  [picker.urls]; that flag also brings in shared/js/url-import.js, the matching
+  stylesheet, and the img-src this page needs, so none of the four can be
+  present without the others.
+
+  The warning is not small print and is not collapsed away with the rest: this
+  is the only feature on the site that contacts anything, and a visitor who
+  opens this panel is owed the reason in full before they use it.
+-->
+<details id="url-panel" class="url-panel">
+  <summary>{{ tool.picker.urls.summary }}</summary>
+  <div class="url-body">
+    <p class="url-warning">
+      <strong>This is the one feature that touches the network.</strong>
+      Fetching an address tells that server your IP address and which {{ tool.picker.urls.noun }} you
+      asked for. The {{ tool.picker.urls.noun }} is copied to local memory on arrival and never
+      requested again. Nothing of yours is ever sent out &mdash; this page still
+      cannot make an upload of any kind.
+    </p>
+    <label for="url-input">One address per line</label>
+    <textarea id="url-input" rows="3" spellcheck="false"
+              placeholder="https://example.com/photo1.jpg&#10;https://example.com/photo2.jpg"></textarea>
+    <div class="url-actions">
+      <button type="button" id="fetch-urls" class="primary">{{ tool.picker.urls.button }}</button>
+      <span id="url-status" class="url-status" role="status" aria-live="polite"></span>
+    </div>
+    <p class="url-note">
+      The server must permit cross-origin use of its {{ tool.picker.urls.noun }}s. Many sites do not,
+      and those will fail with a message rather than silently breaking later.
+    </p>
+  </div>
+</details>
